FinGuru/ai-engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
langchain==0.1.5
langchain-community==0.0.16
langgraph==0.0.20
weaviate-client==3.26.2
sentence-transformers==2.3.1
httpx==0.26.0
pydantic==2.5.3
python-dotenv==1.0.0
